Add CreateFolderCommand
